整理工程不同文件，并归类到文件夹
