+++ .gitignore
@@ -9,5 +9,4 @@
 /Setup/Release
 
 /bak
-/icon
