+++ Properties/AssemblyInfo.cs
@@ -33,6 +33,6 @@ using System.Security;
 //可以指定所有这些值，也可以使用“生成号”和“修订号”的默认值，
 // 方法是按如下所示使用“*”: :
 // [assembly: AssemblyVersion("1.0.*")]
-[assembly: AssemblyVersion("1.3.2.43")]
-[assembly: AssemblyFileVersion("1.3.2.43")]
+[assembly: AssemblyVersion("1.3.3.52")]
+[assembly: AssemblyFileVersion("1.3.3.52")]
