+++ Setup/Setup.vdproj
@@ -553,15 +553,15 @@
         {
         "Name" = "8:Microsoft Visual Studio"
         "ProductName" = "8:mText"
-        "ProductCode" = "8:{4E3659EE-01BD-4F9F-8D8D-E5E84D77BB85}"
-        "PackageCode" = "8:{3BBCF50A-A000-43B3-9A32-B0F4C4F44833}"
+        "ProductCode" = "8:{FF9BBF35-0966-4A17-B31E-4B262D1F1CA5}"
+        "PackageCode" = "8:{20AC98A8-20B9-440B-8980-8E20C3925E58}"
         "UpgradeCode" = "8:{528F43AC-48EB-4B58-85C7-9E6A1FACEED8}"
         "AspNetVersion" = "8:"
         "RestartWWWService" = "11:FALSE"
         "RemovePreviousVersions" = "11:TRUE"
         "DetectNewerInstalledVersion" = "11:TRUE"
         "InstallAllUsers" = "11:FALSE"
-        "ProductVersion" = "8:1.3.2"
+        "ProductVersion" = "8:1.3.3"
         "Manufacturer" = "8:x7890"
         "ARPHELPTELEPHONE" = "8:"
         "ARPHELPLINK" = "8:"
@@ -569,7 +569,7 @@
         "Subject" = "8:"
         "ARPCONTACT" = "8:x7890"
         "Keywords" = "8:"
-        "ARPCOMMENTS" = "8:mText v1.3.2"
+        "ARPCOMMENTS" = "8:mText v1.3.3"
         "ARPURLINFOABOUT" = "8:"
         "ARPPRODUCTICON" = "8:"
         "ARPIconIndex" = "3:0"
