+++ mText.csproj
@@ -179,26 +179,32 @@
     can be found.
   -->
   <ItemGroup>
-    <Compile Include="FormCountSettings.cs">
+    <Compile Include="Forms\CountSettingsForm.cs">
       <SubType>Form</SubType>
     </Compile>
-    <Compile Include="FormCountSettings.Designer.cs">
-      <DependentUpon>FormCountSettings.cs</DependentUpon>
+    <Compile Include="Forms\CountSettingsForm.Designer.cs">
+      <DependentUpon>CountSettingsForm.cs</DependentUpon>
     </Compile>
-    <Compile Include="FormVersion.cs">
+    <Compile Include="Forms\VersionForm.cs">
       <SubType>Form</SubType>
     </Compile>
-    <Compile Include="FormVersion.Designer.cs">
-      <DependentUpon>FormVersion.cs</DependentUpon>
+    <Compile Include="Forms\VersionForm.Designer.cs">
+      <DependentUpon>VersionForm.cs</DependentUpon>
+    </Compile>
+    <Compile Include="Ribbon\gAbout.cs">
+      <SubType>Component</SubType>
+    </Compile>
+    <Compile Include="Ribbon\gBody.cs">
+      <SubType>Component</SubType>
     </Compile>
     <Compile Include="Properties\AssemblyInfo.cs">
       <SubType>Code</SubType>
     </Compile>
-    <EmbeddedResource Include="FormCountSettings.resx">
-      <DependentUpon>FormCountSettings.cs</DependentUpon>
+    <EmbeddedResource Include="Forms\CountSettingsForm.resx">
+      <DependentUpon>CountSettingsForm.cs</DependentUpon>
     </EmbeddedResource>
-    <EmbeddedResource Include="FormVersion.resx">
-      <DependentUpon>FormVersion.cs</DependentUpon>
+    <EmbeddedResource Include="Forms\VersionForm.resx">
+      <DependentUpon>VersionForm.cs</DependentUpon>
     </EmbeddedResource>
     <EmbeddedResource Include="Properties\Resources.resx">
       <Generator>ResXFileCodeGenerator</Generator>
@@ -210,7 +216,7 @@
       <DependentUpon>Resources.resx</DependentUpon>
       <DesignTime>True</DesignTime>
     </Compile>
-    <EmbeddedResource Include="RibbonMain.resx">
+    <EmbeddedResource Include="Ribbon\RibbonMain.resx">
       <DependentUpon>RibbonMain.cs</DependentUpon>
     </EmbeddedResource>
     <None Include="mText_TemporaryKey.pfx" />
@@ -222,14 +228,14 @@
       <AutoGen>True</AutoGen>
       <DependentUpon>Settings.settings</DependentUpon>
     </Compile>
-    <Compile Include="RegPref.cs" />
-    <Compile Include="RibbonMain.cs">
+    <Compile Include="Utils\RegPref.cs" />
+    <Compile Include="Ribbon\RibbonMain.cs">
       <SubType>Component</SubType>
     </Compile>
-    <Compile Include="RibbonMain.Designer.cs">
+    <Compile Include="Ribbon\RibbonMain.Designer.cs">
       <DependentUpon>RibbonMain.cs</DependentUpon>
     </Compile>
-    <Compile Include="ShortcutHookHelper.cs" />
+    <Compile Include="Utils\ShortcutHookHelper.cs" />
     <Compile Include="ThisAddIn.cs">
       <SubType>Code</SubType>
     </Compile>
@@ -241,9 +247,7 @@
     </Compile>
     <AppDesigner Include="Properties\" />
   </ItemGroup>
-  <ItemGroup>
-    <Folder Include="icon\" />
-  </ItemGroup>
+  <ItemGroup />
   <PropertyGroup>
     <VisualStudioVersion Condition="'$(VisualStudioVersion)' == ''">10.0</VisualStudioVersion>
     <VSToolsPath Condition="'$(VSToolsPath)' == ''">$(MSBuildExtensionsPath32)\Microsoft\VisualStudio\v$(VisualStudioVersion)</VSToolsPath>
@@ -270,7 +274,7 @@
           <HostItem Name="ThisAddIn" Code="ThisAddIn.cs" CanonicalName="AddIn" PublishedHash="200B0A089086D201D6B1D6C96257C7B77FBED100" CanActivate="false" IconIndex="1" Blueprint="ThisAddIn.Designer.xml" GeneratedCode="ThisAddIn.Designer.cs" />
         </Host>
       </FlavorProperties>
-      <UserProperties BuildVersion_UseGlobalSettings="True" BuildVersion_StartDate="2000/1/1" />
+      <UserProperties BuildVersion_StartDate="2000/1/1" BuildVersion_UseGlobalSettings="True" />
     </VisualStudio>
   </ProjectExtensions>
 </Project>
